requirements.txt
altgraph==0.17
asgiref==3.3.1
auto-py-to-exe==2.7.11
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.4
click==7.1.2
Django==3.1.4
django-widget-tweaks==1.4.8
Eel==0.12.4
Flask==1.1.2
flaskwebgui==0.1.13
future==0.18.2
gevent==20.12.1
gevent-websocket==0.10.1
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pefile==2019.4.18
Pillow==8.0.1
psutil==5.8.0
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pytz==2020.5
pywin32-ctypes==0.2.0
sqlparse==0.4.1
Werkzeug==1.0.1
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.2.0
